requirements.txt
fastapi
uvicorn
ollama
python-dotenv
requests
python-multipart
python-jose[cryptography]
passlib[bcrypt]

